divide app files
