+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## 0.1.6
+* 出力ファイルの構造を指定する `--bundle` オプションを追加
+  * コンテンツのassetsやglobalScriptsをhtmlに埋め込むか指定できるようになりました。
 ## 0.1.5
 * 出力ファイルの構造を変更
   * 大文字・小文字違いのアセットを利用するコンテンツが正常に動作しない問題を修正しました。
