+++ templates/bundle-index.ect
@@ -1,0 +1,44 @@
+<!doctype html>
+<html>
+<head>
+<meta charset="utf-8">
+<title>game</title>
+<script>
+<% for preloadScript in @preloadScripts : %>
+	<%- preloadScript %>
+<% end %>
+</script>
+
+<script>
+window.g = require("@akashic/akashic-engine");
+</script>
+<script>
+if (! ("gLocalAssetContainer" in window)) {
+	window.gLocalAssetContainer = {};
+}
+
+<% for asset in @assets : %>
+	<% if asset.type is "script" : %>
+		window.gLocalAssetContainer["<%- asset.name %>"] = function(g) {
+			<%- asset.code %>
+		};
+	<% else if asset.type is "text" : %>
+		window.gLocalAssetContainer["<%- asset.name %>"] = "<%- asset.code %>";
+	<% end %>
+<% end %>
+</script>
+
+<script>
+<% for postloadScript in @postloadScripts : %>
+	<%- postloadScript %>
+<% end %>
+</script>
+
+<link rel="stylesheet" type="text/css" href="css/style.css">
+
+</head>
+<body>
+<div id="container">
+</div>
+</body>
+</html>
