--- templates/index.ect
@@ -1,66 +1,0 @@
-<!doctype html>
-<html>
-<head>
-<meta charset="utf-8">
-<title>game</title>
-
-<% if @isBundle : %>
-	<script>
-	<% for preloadScript in @preloadScripts : %>
-		<%- preloadScript %>
-	<% end %>
-	</script>
-<% else : %>
-	<script src="./js/akashic-engine.strip.js"></script>
-	<script src="./js/game-driver.strip.js"></script>
-	<script src="./js/pdi-browser.strip.js"></script>
-<% end %>
-
-<% if @isBundle : %>
-	<script>
-	window.g = require("@akashic/akashic-engine");
-	</script>
-	<script>
-	if (! ("gLocalAssetContainer" in window)) {
-		window.gLocalAssetContainer = {};
-	}
-
-	<% for asset in @assets : %>
-		<% if asset.type is "script" : %>
-			window.gLocalAssetContainer["<%- asset.name %>"] = function(g) {
-				<%- asset.code %>
-			};
-		<% else if asset.type is "text" : %>
-			window.gLocalAssetContainer["<%- asset.name %>"] = "<%- asset.code %>";
-		<% end %>
-	<% end %>
-	</script>
-<% else : %>
-	<script src="./js/initGlobals.js"></script>
-	<% for asset in @assets : %>
-		<script src="<%- asset.name %>"></script>
-	<% end %>
-<% end %>
-
-<% if @isBundle : %>
-	<script>
-	<% for postloadScript in @postloadScripts : %>
-		<%- postloadScript %>
-	<% end %>
-	</script>
-<% else : %>
-	<script src="./js/LocalScriptAsset.js"></script>
-	<script src="./js/LocalTextAsset.js"></script>
-	<script src="./js/game-storage.strip.js"></script>
-	<script src="./js/logger.js"></script>
-	<script src="./js/sandbox.js"></script>
-<% end %>
-
-<link rel="stylesheet" type="text/css" href="css/style.css">
-
-</head>
-<body>
-<div id="container">
-</div>
-</body>
-</html>
