+++ templates/no-bundle-index.ect
@@ -1,0 +1,25 @@
+<!doctype html>
+<html>
+<head>
+<meta charset="utf-8">
+<title>game</title>
+<script src="./js/akashic-engine.strip.js"></script>
+<script src="./js/game-driver.strip.js"></script>
+<script src="./js/pdi-browser.strip.js"></script>
+<script src="./js/initGlobals.js"></script>
+<script src="./js/LocalScriptAsset.js"></script>
+<script src="./js/LocalTextAsset.js"></script>
+<script src="./js/game-storage.strip.js"></script>
+<script src="./js/logger.js"></script>
+<link rel="stylesheet" type="text/css" href="css/style.css">
+
+<% for asset in @assets : %>
+<script src="<%- asset %>"></script>
+<% end %>
+<script src="./js/sandbox.js"></script>
+</head>
+<body>
+<div id="container">
+</div>
+</body>
+</html>
